requirements.txt
Babel==2.5.1
boto3==1.4.8
botocore==1.8.3
dj-database-url==0.5.0
Django==2.0
django-hashid-field==1.2.3
django-heroku==0.3.1
django-money==0.14.4
django-paypal==1.0.0
django-mathfilters==0.4.0
django-phonenumber-field==1.3.0
django-star-ratings==0.8.0
django-storages==1.6.5
django-storages==1.6.5
django-twilio==0.10.0
django-webpush==0.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
hashids==1.2.0
jmespath==0.9.3
olefile==0.44
phonenumberslite==8.8.7
Pillow==6.0.0
psycopg2-binary==2.8.2
psycopg2==2.7.7
py-moneyed==0.8.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
s3transfer==0.1.11
six==1.11.0
stripe==2.27.0
waitress==1.3.0
whitenoise==4.1.2
